Created a function for initializing the PlaygroundLogger framework.
This is `initializePlaygroundLogger()`, which sets up the `_playgroundPrintHook` function with the standard library and the various `$builtin` functions with PlaygroundRuntime.
This commit includes creating a `printHook` function, but as with the other logger entrypoints, it currently just calls `fatalError`.

As part of this, began passing `-Xfrontend -debugger-support` when building PlaygroundLogger so it can use dollar identifiers.