Added an empty umbrella header to the PlaygroundSupport framework.
SourceKit currently is only able to provide code completion in `import` statements for Clang modules.
As a result, to ensure that PlaygroundSupport shows up in code completion, include an empty umbrella header.
This causes PlaygroundSupport.framework to provide a Clang module in addition to the Swift module.

As part of this commit, also fixed a handful of build settings in the PlaygroundSupport_* targets.