[PlaygroundLogger] Unwrap optionality when generating log entries.
The original implementation of PlaygroundLogger treated Optional.some as effectively not existing; it would always unwrap the optionality and present the payload as if it were top-level.
This doesn't quite do the same thing (the type name will be Optional<T>, not T) but it preserves the main behavior.
This fixes testOptionalGetsStripped().