Better handle some cases where PlaygroundLogger previously failed to encode empty images.
If a `NSView`, `NSImage`, or `UIImage` was zero-sized (e.g. 0x0), PlaygroundLogger would fail to encode this because it couldn't get a bitmap representation for the image.
Instead, handle this case specially, and instead of failing, send zero bytes of PNG data. Xcode renders this as an empty image, which is the expectation for this case.
This commit includes tests that empty views and images on all platforms generate image representations which can be encoded without throwing any errors.

This is an improvement over the legacy PlaygroundLogger implementation, which would render these plus empty `UIView` instances as if they were the string "empty image".

This addresses <rdar://problem/40207604>.