[PlaygroundLogger] Added tests for recursive logging.
Added tests for both the new and legacy entrypoints to ensure that recursive logging does not occur.
In both cases, a struct whose implementation of `CustomStringConvertible` logs `self` is logged.
Without the changes in the previous commit, this causes infinite recursion.
With those changes, these tests pass as expected.

This is for <rdar://problem/41460357> / SR-8349.